+++ internal/controller/operator/factory/vlcluster/vlcluster_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
+	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
@@ -45,7 +45,7 @@ func TestCreateOrUpdate(t *testing.T) {
 						return
 					case <-tc.C:
 						if err := cb(); err != nil {
-							if errors.IsNotFound(err) {
+							if k8serrors.IsNotFound(err) {
 								continue
 							}
 							t.Errorf("callback error: %s", err)