+++ internal/controller/operator/factory/vmalertmanager/alertmanager_test.go
@@ -7,9 +7,10 @@ import (
 	"time"
 
 	"github.com/go-test/deep"
+	"github.com/stretchr/testify/assert"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
+	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -222,7 +223,7 @@ func TestCreateOrUpdateAlertManager(t *testing.T) {
 					case <-tc.C:
 						var got appsv1.StatefulSet
 						if err := fclient.Get(ctx, types.NamespacedName{Namespace: tt.args.cr.Namespace, Name: tt.args.cr.PrefixedName()}, &got); err != nil {
-							if !errors.IsNotFound(err) {
+							if !k8serrors.IsNotFound(err) {
 								t.Errorf("cannot get statefulset for alertmanager: %s", err)
 								return
 							}
@@ -348,6 +349,43 @@ func Test_createDefaultAMConfig(t *testing.T) {
 				},
 			},
 		},
+		{
+			name: "with utf-8",
+			args: args{
+				ctx: context.TODO(),
+				cr: &vmv1beta1.VMAlertmanager{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "test-am",
+						Namespace: "default",
+					},
+					Spec: vmv1beta1.VMAlertmanagerSpec{
+						ConfigSecret:            "some-name",
+						ConfigRawYaml:           "global: {}",
+						ConfigSelector:          &metav1.LabelSelector{},
+						ConfigNamespaceSelector: &metav1.LabelSelector{},
+					},
+				},
+			},
+			predefinedObjects: []runtime.Object{
+				&vmv1beta1.VMAlertmanagerConfig{
+					ObjectMeta: metav1.ObjectMeta{
+						Name:      "some",
+						Namespace: "default",
+					},
+					Spec: vmv1beta1.VMAlertmanagerConfigSpec{
+						Route: &vmv1beta1.Route{Receiver: "base", Matchers: []string{`"baf"="daf"`}},
+						Receivers: []vmv1beta1.Receiver{
+							{
+								Name: "base",
+								WebhookConfigs: []vmv1beta1.WebhookConfig{
+									{URL: ptr.To("http://some-url")},
+								},
+							},
+						},
+					},
+				},
+			},
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -363,11 +401,20 @@ func Test_createDefaultAMConfig(t *testing.T) {
 
 			err := fclient.Get(tt.args.ctx, types.NamespacedName{Namespace: tt.args.cr.Namespace, Name: secretName}, &createdSecret)
 			if err != nil {
-				if errors.IsNotFound(err) && tt.secretMustBeMissing {
+				if k8serrors.IsNotFound(err) && tt.secretMustBeMissing {
 					return
 				}
 				t.Fatalf("config for alertmanager not exist, err: %v", err)
 			}
+
+			var amcfgs vmv1beta1.VMAlertmanagerConfigList
+			assert.Nil(t, fclient.List(tt.args.ctx, &amcfgs))
+			for _, amcfg := range amcfgs.Items {
+				if amcfg.Status.UpdateStatus != vmv1beta1.UpdateStatusOperational {
+					t.Fatalf("unexpected non-operational status: %s for amcfg: %s reason: %s ", amcfg.Status.UpdateStatus, amcfg.Name, amcfg.Status.Reason)
+				}
+			}
+
 		})
 	}
 }