+++ docs/CHANGELOG.md
@@ -15,7 +15,15 @@ aliases:
 
 * FEATURE: [vmanomaly](https://docs.victoriametrics.com/anomaly-detection/): update shards concurrently. See [#1434](https://github.com/VictoriaMetrics/operator/issues/1434).
 
+## [v0.60.2](https://github.com/VictoriaMetrics/operator/releases/tag/v0.60.2)
+
+**Release date:** 27 June 2025
+
 * BUGFIX: [vmoperator](https://docs.victoriametrics.com/operator/): clear `status.reason` when transitioning to `expanding` state. This makes less confusing as resource in `expanding` state is no longer affected by an error displayed in `status.reason`. See [#1426](https://github.com/VictoriaMetrics/operator/issues/1426).
+* BUGFIX: [vmalert](https://docs.victoriametrics.com/operator/resources/vmalert/): properly load TLS secrets into VMAlert secret. See [#1437](https://github.com/VictoriaMetrics/operator/issues/1437).
+
+* FEATURE: [vmalertmanagerconfig](https://docs.victoriametrics.com/operator/resources/vmalertmanagerconfig/): support `utf-8` symbols at `matchers`. See this [1446](https://github.com/VictoriaMetrics/operator/issues/1446) issue for details.
+* FEATURE: [vmalertmanager](https://docs.victoriametrics.com/operator/resources/vmalertmanager/): enable `utf-8` support at labels and matchers by default for v0.28.+ versions of alertmanager. See this [1446](https://github.com/VictoriaMetrics/operator/issues/1446) issue for details.
 
 ## [v0.60.1](https://github.com/VictoriaMetrics/operator/releases/tag/v0.60.1)